Added campfire saving

+++ src/main/java/net/rubyhorizon/campfires/campfire/IndicativeCampfire.java
@@ -39,7 +39,7 @@ public class IndicativeCampfire {
         }
     }
 
-    public IndicativeCampfire(Block block, int burningTimeMillis) {
+    public IndicativeCampfire(Block block, long burningTimeMillis) {
         this.location = block.getLocation().toCenterLocation();
         this.burningTimeMillis = burningTimeMillis;
         this.campfireType = Type.getByMaterial(block.getType());
